# backend-java/src/main/java/com/example/calorie/meal/MealType.java
package com.example.calorie.meal;

/**
 * 食事区分。列挙の順序がそのまま画面と集計の表示順になる。
 */
public enum MealType {
    BREAKFAST,
    LUNCH,
    DINNER,
    SNACK
}